Starting to work on in app billing. Coins buttons are now dynamic

+++ GalCon-Client/GalCon-Client/src/com/xxx/galcon/UIConnectionWrapper.java
@@ -57,7 +57,7 @@ public class UIConnectionWrapper {
 
 	}
 	
-	public static void addCoins(UIConnectionResultCallback<Player> callback, String playerHandle, Long numCoins, Long usedCoins){
+	public static void addCoins(UIConnectionResultCallback<Player> callback, String playerHandle, int numCoins, Long usedCoins){
 		try{
 			gameAction.addCoins(callback, playerHandle, numCoins, usedCoins);
 		}catch (ConnectionException e){
